+++ src/core_codemods/replace_flask_send_file.py
@@ -7,8 +7,8 @@ from codemodder.codemods.utils_mixin import NameAndAncestorResolutionMixin
 from codemodder.utils.utils import positional_to_keyword
 
 
-class FlaskSendFilePathParameterization(BaseCodemod, NameAndAncestorResolutionMixin):
-    NAME = "flask-send-file-path-parameterization"
+class ReplaceFlaskSendFile(BaseCodemod, NameAndAncestorResolutionMixin):
+    NAME = "replace-flask-send-file"
     SUMMARY = "Replace unsafe usage of `flask.send_file`"
     DESCRIPTION = SUMMARY
     REVIEW_GUIDANCE = ReviewGuidance.MERGE_WITHOUT_REVIEW
